tests: add tests to cover api changes
